Genericized ManufactoryRecipeSerializer away from GrinderRecipe
